+++ pg.go
@@ -35,8 +35,14 @@ func Scan(values ...interface{}) orm.ColumnScanner {
 	return orm.Scan(values...)
 }
 
-// Safe replaces any placeholders found in the query.
-func Safe(query string, params ...interface{}) types.ValueAppender {
+// Safe represents a safe SQL query.
+type Safe = types.Safe
+
+// Ident represents a SQL identifier, e.g. table or column name.
+type Ident = types.Ident
+
+// SafeQuery replaces any placeholders found in the query.
+func SafeQuery(query string, params ...interface{}) types.ValueAppender {
 	return orm.Safe(query, params...)
 }
 
@@ -45,15 +51,9 @@ var qWarn sync.Once
 // DEPRECATED. Use Safe instead.
 func Q(query string, params ...interface{}) types.ValueAppender {
 	qWarn.Do(func() {
-		internal.Logger.Printf("DEPRECATED: pg.Q is replaced with pg.Safe")
+		internal.Logger.Printf("DEPRECATED: pg.Q is replaced with pg.SafeQuery")
 	})
-	return Safe(query, params...)
-}
-
-// Ident quotes a SQL identifier such as a table or column name replacing any
-// placeholders found in the field.
-func Ident(field string) types.ValueAppender {
-	return types.Ident(field)
+	return SafeQuery(query, params...)
 }
 
 var fWarn sync.Once